pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastembed>=0.7.4",
    "google-generativeai>=0.8.5",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.3",
    "ollama>=0.6.1",
    "pandas>=2.3.3",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "sec-edgar-downloader>=5.0.3",
    "sqlalchemy>=2.0.45",
    "unstructured>=0.18.21",
]
